+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `TimeoutObserver`.
 - Fixed KVO for the `iSCancelled` property in `AdvancedOperation`.
 - A `GroupOperation` can be cancelled correctly even if it's not yet started.
+- The propery `failed` in `AdvancedOperation` has been renamed `hasErrors` to better reflect its intent.
 - Bugfixes.
 
 ### 1.1.0